refactor beacon sync and tailor by functionality into 5 sub-files
