+++ src/beacon_chain/blocks/heal.rs
@@ -62,13 +62,12 @@ pub async fn heal_block_hashes() {
         first_slot.0
     )
     .fetch(&db_pool);
-    /**
-        `work_todo` is a query that counts how many rows in the `beacon_blocks` table need to be processed.
-        specifically where `block_hash` is NULL and the `slot` is greater than or equal to `first_slot.0`.
-        This count is used to track the total number of blocks that require "healing"(i.e., updating the block hash).
-        We use this count to initialize the progress tracker, ensuring the healing process can report progress as it
-        processes each block.
-    */
+
+    // `work_todo` is a query that counts how many rows in the `beacon_blocks` table need to be processed.
+    // specifically where `block_hash` is NULL and the `slot` is greater than or equal to `first_slot.0`.
+    // This count is used to track the total number of blocks that require "healing"(i.e., updating the block hash).
+    // We use this count to initialize the progress tracker, ensuring the healing process can report progress as it
+    //processes each block.
     let mut progress =
         Progress::new("heal-block-hashes", work_todo.count.try_into().unwrap());
