+++ src/beacon_chain/syncer/cache_refresh.rs
@@ -1,0 +1,6 @@
+use sqlx::PgPool;
+
+pub async fn update_deferrable_analysis(db_pool: &PgPool) -> anyhow::Result<()> {
+    // todo : refresh update cache, but now we haven't implement this yet
+    Ok(())
+}
